Split bit interleaving implementations and tests
Split each implementation and corresponding unit test to their own
files.

Split benchmarks to 64 bit and 128 bit test cases.
