+++ src/benchmark/CMakeLists.txt
@@ -1,1 +1,2 @@
-Benchmark(BitInterleavingBenchmark)
+Benchmark(BitInterleaving128bitBenchmark)
+Benchmark(BitInterleaving64bitBenchmark)
